hybrid memfield
