+++ .project/.idea/libraries/grpc_core_1_12_0_ijar_4f82d076.xml
@@ -1,7 +1,7 @@
 <component name="libraryTable">
   <library name="grpc-core-1.12.0-ijar_4f82d076">
     <CLASSES>
-      <root url="jar://$PROJECT_DIR$/../bazel-out/x64_windows-fastbuild/genfiles/external/org_pubref_rules_protobuf/java/_ijar/grpc_compiletime_deps/external/io_grpc_grpc_core/jar/grpc-core-1.12.0-ijar.jar!/" />
+      <root url="jar://$USER_HOME$/_bazel_thboo/qegdutd6/execroot/__main__/bazel-out/x64_windows-fastbuild/genfiles/external/org_pubref_rules_protobuf/java/_ijar/grpc_compiletime_deps/external/io_grpc_grpc_core/jar/grpc-core-1.12.0-ijar.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
