+++ .project/.idea/libraries/guava_25_0_jre_ijar_dd03fe95.xml
@@ -1,7 +1,7 @@
 <component name="libraryTable">
   <library name="guava-25.0-jre-ijar_dd03fe95">
     <CLASSES>
-      <root url="jar://$PROJECT_DIR$/../bazel-out/x64_windows-fastbuild/genfiles/external/org_pubref_rules_protobuf/java/_ijar/grpc_compiletime_deps/external/com_google_guava_guava/jar/guava-25.0-jre-ijar.jar!/" />
+      <root url="jar://$USER_HOME$/_bazel_thboo/qegdutd6/execroot/__main__/bazel-out/x64_windows-fastbuild/genfiles/external/org_pubref_rules_protobuf/java/_ijar/grpc_compiletime_deps/external/com_google_guava_guava/jar/guava-25.0-jre-ijar.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
