+++ .project/.idea/libraries/jsr305_3_0_0_ijar_12061eea.xml
@@ -1,7 +1,7 @@
 <component name="libraryTable">
   <library name="jsr305-3.0.0-ijar_12061eea">
     <CLASSES>
-      <root url="jar://$PROJECT_DIR$/../bazel-out/x64_windows-fastbuild/genfiles/external/org_pubref_rules_protobuf/java/_ijar/grpc_compiletime_deps/external/com_google_code_findbugs_jsr305/jar/jsr305-3.0.0-ijar.jar!/" />
+      <root url="jar://$USER_HOME$/_bazel_thboo/qegdutd6/execroot/__main__/bazel-out/x64_windows-fastbuild/genfiles/external/org_pubref_rules_protobuf/java/_ijar/grpc_compiletime_deps/external/com_google_code_findbugs_jsr305/jar/jsr305-3.0.0-ijar.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
