+++ .project/.idea/libraries/junit_4_12_ijar_b66bc329.xml
@@ -1,7 +1,7 @@
 <component name="libraryTable">
   <library name="junit-4.12-ijar_b66bc329">
     <CLASSES>
-      <root url="jar://$PROJECT_DIR$/../bazel-out/x64_windows-fastbuild/genfiles/external/junit_junit/jar/_ijar/jar/external/junit_junit/jar/junit-4.12-ijar.jar!/" />
+      <root url="jar://$USER_HOME$/_bazel_thboo/qegdutd6/execroot/__main__/bazel-out/x64_windows-fastbuild/genfiles/external/junit_junit/jar/_ijar/jar/external/junit_junit/jar/junit-4.12-ijar.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
