+++ .project/.idea/libraries/rocksdbjni_5_11_3_ijar_b055ca27.xml
@@ -1,7 +1,7 @@
 <component name="libraryTable">
   <library name="rocksdbjni-5.11.3-ijar_b055ca27">
     <CLASSES>
-      <root url="jar://$PROJECT_DIR$/../bazel-out/x64_windows-fastbuild/genfiles/external/org_rocksdb_rocksdbjni/jar/_ijar/jar/external/org_rocksdb_rocksdbjni/jar/rocksdbjni-5.11.3-ijar.jar!/" />
+      <root url="jar://$USER_HOME$/_bazel_thboo/qegdutd6/execroot/__main__/bazel-out/x64_windows-fastbuild/genfiles/external/org_rocksdb_rocksdbjni/jar/_ijar/jar/external/org_rocksdb_rocksdbjni/jar/rocksdbjni-5.11.3-ijar.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
