+++ miner/src/MrPlow.java
@@ -4,7 +4,6 @@ import com.google.protobuf.ByteString;
 import duckutil.Config;
 import duckutil.ConfigFile;
 import duckutil.TimeRecord;
-import duckutil.TimeRecordAuto;
 import io.grpc.ManagedChannel;
 import io.grpc.ManagedChannelBuilder;
 import io.grpc.stub.StreamObserver;
@@ -13,8 +12,6 @@ import snowblossom.proto.*;
 import snowblossom.mining.proto.*;
 import snowblossom.proto.UserServiceGrpc.UserServiceBlockingStub;
 import snowblossom.proto.UserServiceGrpc.UserServiceStub;
-import snowblossom.lib.trie.HashUtils;
-import snowblossom.lib.SnowMerkleProof;
 import snowblossom.lib.db.DB;
 import snowblossom.lib.db.lobstack.LobstackDB;
 import snowblossom.lib.db.rocksdb.JRocksDB;
@@ -23,15 +20,8 @@ import io.grpc.Server;
 import io.grpc.ServerBuilder;
 
 
-import java.io.File;
-import java.io.FileInputStream;
 import java.math.BigInteger;
-import java.nio.ByteBuffer;
-import java.security.MessageDigest;
 import java.text.DecimalFormat;
-import java.util.Collections;
-import java.util.LinkedList;
-import java.util.Random;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.logging.Level;
 import java.util.logging.Logger;
