+++ miner/src/Sweeper.java
@@ -1,6 +1,5 @@
 package snowblossom.miner;
 
-import snowblossom.lib.SnowMerkleProof;
 import snowblossom.proto.Block;
 
 import java.nio.ByteBuffer;
